+++ ricaun.RevitTest.TestAdapter/TestExecutor.cs
@@ -6,6 +6,7 @@ using ricaun.RevitTest.TestAdapter.Services;
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Threading;
 using System.Threading.Tasks;
 
 namespace ricaun.RevitTest.TestAdapter
@@ -61,10 +62,48 @@ namespace ricaun.RevitTest.TestAdapter
             task.GetAwaiter().GetResult();
         }
 
-        public void Cancel()
+        #region Cancel
+        /// <summary>
+        /// Force to cancel `RunTests` if process exit
+        /// </summary>
+        private void InitializeCancel()
         {
+            AppDomain.CurrentDomain.ProcessExit += CurrentDomain_ProcessExit;
         }
 
+        private void FinishCancel()
+        {
+            AppDomain.CurrentDomain.ProcessExit -= CurrentDomain_ProcessExit;
+        }
+
+        private void CurrentDomain_ProcessExit(object sender, EventArgs e)
+        {
+            Cancel();
+        }
+
+        /// <summary>
+        /// Cancel force to kill process.
+        /// </summary>
+        public void Cancel()
+        {
+            FinishCancel();
+            try
+            {
+                foreach (var process in ricaun.RevitTest.Command.Process.ProcessStart.GetProcesses())
+                {
+#if NET
+                    process.Kill(true);
+#else
+                    process.Kill();
+#endif
+                }
+            }
+            catch (Exception ex)
+            {
+                System.IO.File.WriteAllText($"{this.GetType().Assembly.GetName().Name}.Exception.txt", ex.ToString());
+            }
+        }
+        #endregion
         private static bool IsLogDebug => System.Diagnostics.Debugger.IsAttached || AdapterSettings.Settings.NUnit.Verbosity > 2;
 
         /// <summary>
@@ -76,6 +115,8 @@ namespace ricaun.RevitTest.TestAdapter
         /// <returns></returns>
         private async Task RunTests(IFrameworkHandle frameworkHandle, string source, List<TestCase> tests = null)
         {
+            InitializeCancel();
+
             tests = tests ?? new List<TestCase>();
 
             AdapterLogger.Logger.Info($"RunTest: {source} [TestCase: {tests.Count}]");
@@ -85,7 +126,7 @@ namespace ricaun.RevitTest.TestAdapter
             }
             MetadataSettings.Create(source);
             EnvironmentSettings.Create();
-            
+
             AdapterLogger.Logger.Info("---------");
             AdapterLogger.Logger.Info($"RevitTestConsole: {AdapterSettings.Settings.NUnit.Application}");
             AdapterLogger.Logger.Info("---------");
@@ -114,12 +155,14 @@ namespace ricaun.RevitTest.TestAdapter
                     AdapterSettings.Settings.NUnit.Open,
                     AdapterSettings.Settings.NUnit.Close,
                     AdapterSettings.Settings.NUnit.Timeout,
-                    AdapterLogger.Logger.Debug, (message) => { if (IsLogDebug) AdapterLogger.Logger.Debug(message); }, AdapterLogger.Logger.Warning, 
+                    AdapterLogger.Logger.Debug, (message) => { if (IsLogDebug) AdapterLogger.Logger.Debug(message); }, AdapterLogger.Logger.Warning,
                     filters);
 
             }
 
             AdapterLogger.Logger.Info("---------");
+
+            FinishCancel();
         }
 
         private static void RecordResultTestModel(IFrameworkHandle frameworkHandle, string source, List<TestCase> tests, TestModel testModel)